DETOUR
I wanted to find a way so that users can easily install this. Thought that I'd come up with an "add-on" GAS script and a "dynamic" GAS script. The first would be what I publish on add-on store that would install the dynamic one for each user  (because everyone needs their own script ID).  Long story short, Google doesn't allow for any of this. I just made the project unnecessarily complicated and the user still needs to do a bunch of manual stuff to get it installed.  Will revert soon (except I made a bunch of other core changes too)